+++ src/keyboard.js
@@ -1,28 +1,26 @@
 "use strict";
 
-import Items from './modules/items.js';
+import PageAction from './modules/page_action.js';
+import Keyboard   from './modules/keyboard.js';
+import Settings   from './modules/settings.js';
+import { KeyboardShortcuts } from './modules/constants.js';
 
 
 // ---------------
 
 
+Settings.init().then( function() {
+  let settings = Settings.get();
+
+  Keyboard.registerShortcut(KeyboardShortcuts.toggle,    settings.keyboardToggle);
+  Keyboard.registerShortcut(KeyboardShortcuts.openPopup, settings.keyboardOpenPopup);
+});
+
+
 browser.commands.onCommand.addListener( (command) => {
-  if( command === "toggle-page-status" ) {
+  if( command === KeyboardShortcuts.toggle ) {
     browser.tabs.query({ active: true, currentWindow: true }).then( ([currentTab]) => {
-      const currentUrl   = currentTab.url;
-      const currentTitle = currentTab.title;
-
-      browser.storage.local.get('items').then( ({ items }) => {
-        const matchingItem = Items.find( items, { url: currentUrl });
-
-        if( matchingItem ) {
-          Items.markAsRead( matchingItem.id );
-        } else {
-          const addItemOptions = { closeTabId: currentTab.id };
-          Items.addItem( currentUrl, currentTitle, addItemOptions );
-        }
-      });
+      PageAction.toggle(currentTab);
     });
   }
 });
-