+++ src/modules/constants.js
@@ -1,5 +1,7 @@
 "use strict";
 
+const consumerKey = '58817-addc87503598b7ed29e5bf72';
+
 const PocketError = {
   GENERIC:      'generic',
   UNREACHABLE:  'unreachable',
@@ -12,8 +14,6 @@ const PocketNotice = {
   ALREADY_IN_LIST: 'already_in_list'
 };
 
-const consumerKey = '58817-addc87503598b7ed29e5bf72';
-
 // https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/button
 const MouseButtons = {
   LEFT:   0,
@@ -21,5 +21,10 @@ const MouseButtons = {
   RIGHT:  2
 };
 
-export { PocketError, PocketNotice, MouseButtons, consumerKey };
+const KeyboardShortcuts = {
+  toggle: 'toggle-page-status',
+  openPopup: '_execute_browser_action'
+};
+
+export { KeyboardShortcuts, PocketError, PocketNotice, MouseButtons, consumerKey };